--- deafrica_waterbodies/waterbodies/polygons/polygons.py
@@ -1,199 +1,0 @@
-import datacube
-import os
-import math
-import logging
-import shapely
-import numpy as np
-import pandas as pd
-import geopandas as gpd
-
-from datacube.utils.geometry import Geometry
-from deafrica_tools.spatial import xr_vectorize
-
-from .helpers import check_wetness_thresholds, setup_output_fp, assign_unique_ids
-from .filter import filter_waterbodies
-
-
-_log = logging.getLogger(__name__)
-
-
-def get_polygons_using_thresholds(
-        input_gdf: gpd.geodataframe.GeoDataFrame,
-        dask_chunks: dict[str, int] = {"x": 3000, "y": 3000, "time": 1},
-        resolution: tuple[int, int] = (-30, 30),
-        output_crs: str = "EPSG:6933",
-        min_valid_observations: int = 128,
-        primary_threshold: float = 0.1,
-        secondary_threshold: float = 0.05,
-                            ):
-    """
-    Generate polygons by thresholding WOfS All Time Summary data.
-
-    Parameters
-    ----------
-    input_gdf : gpd.geodataframe.GeoDataFrame
-        Area of interest GeoDataFrame
-    dask_chunks : dict, optional
-        dask_chunks to use to load WOfS data, by default {"x": 3000, "y": 3000, "time": 1}
-    resolution : tuple[int, int], optional
-        Resolution to use to load WOfS data, by default (-30, 30)
-    output_crs : str, optional
-        CRS to load data and for the output waterbody polygons, by default "EPSG:6933"
-    min_valid_observations : int, optional
-        Threshold to use to mask out pixels based on the number of valid WOfS observations for each pixel, by default 128
-    primary_threshold : float, optional
-        Threshold to use to determine the location of the waterbody polygons, by default 0.1
-    secondary_threshold : float, optional
-        Threshold to use to determine the extent / shape of the waterbodies polygons, by default 0.05
-
-    Returns
-    -------
-    primary_threshold_polygons, secondary_threshold_polygons
-        Polygons generated from thresholding WOfS All Time Summary data
-        using the primary and secondary thresholds.
-
-    """
-    # Check if the wetness thresholds have been set correctly.
-    minimum_wet_thresholds = [secondary_threshold, primary_threshold]
-    _log.info(check_wetness_thresholds(minimum_wet_thresholds))
-
-    # Create a datacube query object.
-    query = dict(
-        dask_chunks=dask_chunks,
-        resolution=resolution,
-        output_crs=output_crs,
-    )
-
-    # Reproject the input_gdf.
-    input_gdf = input_gdf.to_crs(output_crs)
-
-    # Connect to the datacube.
-    dc = datacube.Datacube(app='WaterbodiesPolygons')
-
-    primary_threshold_polygons_list = []
-    secondary_threshold_polygons_list = []
-    for row in input_gdf.itertuples():
-        row_id = row.Index
-        _log.info(f"Generating polygons for row: {row_id}")
-        row_geom = Geometry(geom=row.geometry, crs=input_gdf.crs)
-
-        # Update the datacube query with the row geometry.
-        query.update(geopolygon=row_geom)
-
-        # Load the WOfS All-Time Summary of clear and wet observations.
-        wofs_alltime_summary = dc.load('wofs_ls_summary_alltime',
-                                       **query).squeeze()
-
-        # Set the no-data values to nan.
-        # Masking here is done using the frequency measurement because for multiple
-        # areas NaN values are present in the frequency measurement but the
-        # no data value -999 is not present in the count_clear and
-        # count_wet measurements.
-        # Note: it seems some pixels with NaN values in the frequency measurement
-        # have a value of zero in the count_clear and/or the count_wet measurements.
-        wofs_alltime_summary = wofs_alltime_summary.where(~np.isnan(wofs_alltime_summary.frequency))
-
-        # Mask pixels not observed at least min_valid_observations times.
-        wofs_alltime_summary_valid_clear_count = wofs_alltime_summary.count_clear >= min_valid_observations
-
-        # Generate the polygons.
-        row_polygons = {}
-        for threshold in minimum_wet_thresholds:
-            # Mask any pixels whose frequency of water detection is less than the threshold.
-            wofs_alltime_summary_valid_wetness = wofs_alltime_summary.frequency > threshold
-
-            # Now find pixels that meet both the minimum valid observations
-            # and minimum wet threshold criteria.
-            wofs_alltime_summary_valid = wofs_alltime_summary_valid_wetness.where(wofs_alltime_summary_valid_wetness & wofs_alltime_summary_valid_clear_count)
-
-            # Convert the raster to polygons.
-            # We use a mask of '1' to only generate polygons around values of '1' (not NaNs).
-            polygons_mask = wofs_alltime_summary_valid == 1
-
-            polygons = xr_vectorize(wofs_alltime_summary_valid,
-                                    mask=polygons_mask,
-                                    crs=wofs_alltime_summary.geobox.crs,
-                                    transform=wofs_alltime_summary.geobox.transform)
-
-            # Combine any overlapping polygons.
-            merged_polygon_geoms = shapely.ops.unary_union(polygons['geometry'])
-
-            # Turn the combined multipolygon back into a GeoDataFrame.
-            merged_polygons = gpd.GeoDataFrame(geometry=list(merged_polygon_geoms.geoms))
-
-            # We need to add the crs back onto the GeoDataFrame.
-            merged_polygons.crs = wofs_alltime_summary.geobox.crs
-
-            row_polygons[threshold] = merged_polygons
-
-        # Append the row's waterbody polygons to the list.
-        primary_threshold_polygons_list.append(row_polygons[primary_threshold])
-        secondary_threshold_polygons_list.append(row_polygons[secondary_threshold])
-
-    primary_threshold_polygons = pd.concat(primary_threshold_polygons_list)
-    secondary_threshold_polygons = pd.concat(secondary_threshold_polygons_list)
-
-    return primary_threshold_polygons, secondary_threshold_polygons
-
-
-def get_waterbodies(
-        aoi_gdf: gpd.geodataframe.GeoDataFrame,
-        dask_chunks: dict[str, int] = {"x": 3000, "y": 3000, "time": 1},
-        resolution: tuple[int, int] = (-30, 30),
-        output_crs: str = "EPSG:6933",
-        min_valid_observations: int = 128,
-        primary_threshold: float = 0.1,
-        secondary_threshold: float = 0.05,
-        min_polygon_size: float = 4500,
-        max_polygon_size: float = math.inf,
-        filter_out_ocean_polygons: bool = False,
-        land_sea_mask_fp: str = None,
-        filter_out_major_rivers_polygons: bool = False,
-        major_rivers_mask_fp: str = None,
-        filter_out_urban_polygons: bool = False,
-        urban_mask_fp: str = None,
-        handle_large_polygons: str = "nothing",
-        pp_test_threshold: float = 0.005,
-        ouptut_folder: str = os.getcwd(),
-        output_base_filename: str = "waterbodies",
-        output_file_type: str = "GeoJSON",
-):
-    # Set up some file names for outputs
-    _log.info("Setting up output directory")
-
-    final_output_fp = setup_output_fp(ouptut_folder=ouptut_folder,
-                                      output_base_filename=output_base_filename,
-                                      output_file_type=output_file_type)
-
-    _log.info("Generating the first temporary set of waterbody polygons.")
-    temp_primary, temp_secondary = get_polygons_using_thresholds(input_gdf=aoi_gdf,
-                                                                 dask_chunks=dask_chunks,
-                                                                 resolution=resolution,
-                                                                 output_crs=output_crs,
-                                                                 min_valid_observations=min_valid_observations,
-                                                                 primary_threshold=primary_threshold,
-                                                                 secondary_threshold=secondary_threshold,
-                                                                 )
-    _log.info("Filter waterbodies")
-    filtered_polygons = filter_waterbodies(
-        primary_threshold_polygons=temp_primary,
-        secondary_threshold_polygons=temp_secondary,
-        min_polygon_size=min_polygon_size,
-        max_polygon_size=max_polygon_size,
-        filter_out_ocean_polygons=filter_out_ocean_polygons,
-        land_sea_mask_fp=land_sea_mask_fp,
-        filter_out_major_rivers_polygons=filter_out_major_rivers_polygons,
-        major_rivers_mask_fp=major_rivers_mask_fp,
-        filter_out_urban_polygons=filter_out_urban_polygons,
-        urban_mask_fp=urban_mask_fp,
-        handle_large_polygons=handle_large_polygons,
-        pp_test_threshold=pp_test_threshold,
-    )
-
-    _log.info("Assigning unique ids to each polygon.")
-    filtered_polygons_with_unique_ids = assign_unique_ids(filtered_polygons)
-
-    _log.info("Writing waterbodies to disk...")
-    # Reproject to EPSG:4326
-    filtered_polygons_with_unique_ids = filtered_polygons_with_unique_ids.to_crs("EPSG:4326")
-    filtered_polygons_with_unique_ids.to_file(final_output_fp, driver=output_file_type)
